Ready for use to create a better index.

+++ SparqlForHumans.Core/Utilities/PathExtensions.cs
@@ -4,9 +4,9 @@ namespace SparqlForHumans.Core.Utilities
 {
     public static class PathExtensions
     {
-        public static void DeleteIfExists(this string path)
+        public static void DeleteIfExists(this string path, bool additionalCondition = false)
         {
-            if (Directory.Exists(path))
+            if (Directory.Exists(path) && additionalCondition)
                 Directory.Delete(path, true);
         }
     }
